# docs/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "ApiErrorDocument": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "schema_version": {
            "const": "api-error/v1",
            "default": "api-error/v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorDocument",
        "type": "object"
      },
...
      "ComparisonCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "baseline_run_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$",
            "title": "Baseline Run Id",
            "type": "string"
          },
          "candidate_run_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$",
            "title": "Candidate Run Id",
            "type": "string"
          },
          "dataset_name": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:/-]*$",
            "title": "Dataset Name",
            "type": "string"
          },
          "dataset_revision": {
            "exclusiveMinimum": 0.0,
            "title": "Dataset Revision",
            "type": "integer"
          },
          "spec": {
            "$ref": "#/components/schemas/EvaluationSpecInput"
          }
        },
        "required": [
          "dataset_name",
          "dataset_revision",
          "baseline_run_id",
          "candidate_run_id",
          "spec"
        ],
        "title": "ComparisonCreateRequest",
        "type": "object"
      },
      "ComparisonSubmissionResponse": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReleaseDecisionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "job": {
            "$ref": "#/components/schemas/JobResponse"
          },
          "schema_version": {
            "const": "comparison-submission/v2",
            "default": "comparison-submission/v2",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "job"
        ],
        "title": "ComparisonSubmissionResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ProjectBearer": {
        "bearerFormat": "cpk_<base64url>",
        "description": "Project-bound opaque API credential",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Durable, content-addressed evaluation and release decisions",
    "title": "LLM Evaluation Control Plane",
    "version": "1.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/comparisons": {
      "post": {
        "description": "The API pins stored run evidence and durably enqueues an asynchronous release comparison.",
        "operationId": "submit_release_comparison",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "description": "Configured single-deployment project boundary",
            "in": "header",
            "name": "X-Project-ID",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComparisonCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparisonSubmissionResponse"
                }
              }
            },
            "description": "Terminal replay",
            "headers": {
              "Location": {
                "description": "Canonical path of the durable submission job",
                "schema": {
                  "pattern": "^/v1/jobs/[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$",
                  "type": "string"
                }
              }
            }
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparisonSubmissionResponse"
                }
              }
            },
            "description": "Accepted new or nonterminal job",
            "headers": {
              "Location": {
                "description": "Canonical path of the durable submission job",
                "schema": {
                  "pattern": "^/v1/jobs/[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDocument"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDocument"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDocument"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDocument"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDocument"
                }
              }
            },
            "description": "Immutable conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDocument"
                }
              }
            },
            "description": "Request body too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDocument"
                }
              }
            },
            "description": "Unsupported media type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDocument"
                }
              }
            },
            "description": "Contract validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDocument"
                }
              }
            },
            "description": "Internal service error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDocument"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "ProjectBearer": []
          }
        ],
        "summary": "Submit Comparison",
        "tags": [
          "release decisions"
        ]
      }
    },
...
    "/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "description": "Configured single-deployment project boundary",
            "in": "header",
            "name": "X-Project-ID",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDocument"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDocument"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDocument"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDocument"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDocument"
                }
              }
            },
            "description": "Immutable conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDocument"
                }
              }
            },
            "description": "Request body too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDocument"
                }
              }
            },
            "description": "Unsupported media type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDocument"
                }
              }
            },
            "description": "Contract validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDocument"
                }
              }
            },
            "description": "Internal service error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDocument"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "ProjectBearer": []
          }
        ],
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}